Fix: fix the output of sifted to include changes in filter
